🧹 Create the other section components; tidy files of old stuff
