+++ src/views/Home.vue
@@ -27,208 +27,18 @@
 		</svg>
 
 		<TcHomeWork />
-
-		<!-- <div class="tc-image tc-image-home tc-col-left">
-			<div class="tc-image-wrap">
-				<transition name="opacity">
-					<div 
-						v-if="activeImage"
-						class="tc-image-img"
-						:style="{ 'background-image': activeImage && `url(${activeImage})` }"
-					></div>
-				</transition>
-				<svg class="tc-image-pattern">
-					<defs>
-						<pattern id="dots" width="6.59" height="6.59" patternUnits="userSpaceOnUse" viewBox="0 0 6.59 6.59">
-							<path class="tc-dot" d="M2.11,0A1.33,1.33,0,0,0,3.43,1.29,1.33,1.33,0,0,0,4.75,0Z"/>
-							<path class="tc-dot" d="M2.11,6.59H4.75a1.32,1.32,0,0,0-2.64,0Z"/>
-							<path class="tc-dot" d="M6.59,2h0a1.32,1.32,0,0,0,0,2.64h0Z"/>
-							<path class="tc-dot" d="M0,2H0V4.68H0A1.32,1.32,0,1,0,0,2Z"/>
-						</pattern>
-						<pattern id="dotsSpaced" width="13" height="13" patternUnits="userSpaceOnUse" viewBox="0 0 13 13">
-							<path class="tc-dot" d="M5.32,0A1.33,1.33,0,0,0,6.64,1.29,1.33,1.33,0,0,0,8,0Z"/>
-							<path class="tc-dot" d="M5.32,13H8a1.32,1.32,0,0,0-2.64,0Z"/>
-							<path class="tc-dot" d="M13,5.25h0a1.32,1.32,0,0,0,0,2.64h0Z"/>
-							<path class="tc-dot" d="M0,5.25H0V7.89H0A1.32,1.32,0,0,0,0,5.25Z"/>
-						</pattern>
-					</defs>
-					<rect height="100%" width="100%" fill="url(#dotsSpaced)"></rect>
-				</svg>
-			</div>
-		</div> -->
-		<!-- <div class="tc-content tc-col-right">
-			<section 
-				id="about"
-				class="tc-section"
-			>
-				<h1 class="tc-name">
-					<span class="tc-name-first">Tiff</span>any<br>
-					Choong
-				</h1>
-				<p class="tc-font-2">
-					Front end developer + designer
-				</p>
-
-				<TcCard 
-					header="Hello! 👋"
-					class="my-5"
-				>
-					<p>
-						I’m Tiff. I’m a front end developer and designer who enjoys creating and experimenting in the digital world.
-					</p>
-					<p>
-						I am currently working as a Lead UI Engineer on a digital focused team within a nationally known and recognized banking institution. 
-					</p>
-					<p>
-						My focus is overseeing and developing websites and applications, creating interactive web experiences to surprise and delight customers who craved fun in their banking, and developing and maintaining a design system and component library that made that easy for us.
-					</p>
-				</TcCard>
-
-				<h2>
-					Stack starter pack
-				</h2>
-				<ul class="tc-skill-grid">
-					<li 
-						v-for="skill in skills"
-						:key="skill.label"
-						class="tc-skill-grid-item"
-					>
-						<FontAwesomeIcon 
-							:icon="`fab fa-${skill.icon}`"
-							size="2x"
-							class="tc-skill-grid-icon"
-						/>
-						<span class="tc-skill-grid-text">
-							{{ skill.label }}
-						</span>
-					</li>
-				</ul>
-			</section>
-			<section 
-				id="work"
-				class="tc-section tc-section-bleed"
-			>
-				<header>
-					<h2>
-						Featured work
-					</h2>
-				</header>
-				<ul class="tc-list">
-					<li 
-						v-for="item in works"
-						:key="item.title"
-						class="tc-list-item"
-					>
-						<router-link 
-							class="tc-list-link"
-							:to="`/work/${item.id}`"
-							@mouseover="activeImage = item.images[0]?.src"
-							@mouseout="activeImage = ''"
-						>
-							{{ item.title }}
-						</router-link>
-					</li>
-				</ul>
-			</section>
-			<section 
-				id="media"
-				class="tc-section"
-			>
-				<header>
-					<h2>
-						Featured media
-					</h2>
-				</header>
-				<ul class="tc-list-cards">
-					<li
-						v-for="item in media"
-						:key="item.title"
-						class="tc-list-card"
-					>
-						<a 
-							:href="item.link"
-							class="tc-list-card-link"
-							target="_blank"
-							@mouseover="activeImage = item.image"
-							@mouseout="activeImage = ''"
-						>
-							<span class="mr-2">
-								{{ item.title }}
-								<br>
-								<span class="tc-font-small">
-									{{ item.date }}
-								</span>
-							</span>
-							
-							<TcArrow class="tc-list-card-link-icon" />
-						</a>
-					</li>
-				</ul>
-			</section>
-		</div> -->
+		<TcHomeSkills />
+		<TcHomeMedia />
+		<TcHomeAbout />
 	</div>
 </template>
 
 <script setup lang="ts">
-import { ref } from 'vue';
-import { useWorkStore } from '../stores/work';
-
 import TcHomeSplash from '../components/home/HomeSplash.vue';
 import TcHomeWork from '../components/home/HomeWork.vue';
-
-import TcCard from '../components/Card.vue';
-import TcArrow from '../components/Arrow.vue';
-
-import { library } from '@fortawesome/fontawesome-svg-core';
-import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
-import { faHtml5, faCss3Alt, faJs, faVuejs, faGitAlt, faFigma } from '@fortawesome/free-brands-svg-icons';
-
-library.add(faHtml5, faCss3Alt, faJs, faVuejs, faGitAlt, faFigma);
-
-const store = useWorkStore();
-const { works } = store;
-
-const activeImage = ref('');
-const skills = ref([
-	{
-		icon: 'html5',
-		label: 'HTML'
-	},
-	{
-		icon: 'css3-alt',
-		label: 'CSS'
-	},
-	{
-		icon: 'js',
-		label: 'Javascript'
-	},
-	{
-		icon: 'vuejs',
-		label: 'Vue.js'
-	},
-	{
-		icon: 'git-alt',
-		label: 'Git'
-	},
-	{
-		icon: 'figma',
-		label: 'Figma'
-	}
-]);
-const media = ref([
-	{
-		title: 'Codepen Radio: Episode 359',
-		link: 'https://blog.codepen.io/2022/03/16/359-tiffany-choong/',
-		image: '/img/codepen-radio-01.webp',
-		date: 'March 2022'
-	},
-	{
-		title: 'Net Magazine: Create CSS art',
-		link: 'https://www.creativebloq.com/how-to/create-animated-css-art',
-		image: '/img/netmagazine-01.webp',
-		date: 'July 2019'
-	}
-]);
+import TcHomeSkills from '../components/home/HomeSkills.vue';
+import TcHomeMedia from '../components/home/HomeMedia.vue';
+import TcHomeAbout from '../components/home/HomeAbout.vue';
 </script>
 
 <style lang="scss">
@@ -253,93 +63,6 @@ const media = ref([
 				aspect-ratio: 14/1;
 				background-color: inherit;
 			}
-		}
-	}
-
-
-	// old
-	&-skill-grid {
-		display: grid;
-		gap: var(--spacer-1);
-		grid-template-columns: repeat(auto-fit, 5rem);
-
-		&-text {
-			display: block;
-			margin-top: var(--spacer-1);
-			font-family: var(--font-fam-2);
-			font-size: var(--font-size-sm-2);
-		}
-
-		&-item {
-			display: flex;
-			justify-content: center;
-			flex-direction: column;
-			aspect-ratio: 1/1;
-			border: var(--border);
-			box-shadow: var(--box-shadow-primary);
-			background-color: var(--color-background);
-			text-align: center;
-		}
-	}
-
-	&-list {
-		&-link {
-			display: block;
-			padding: var(--spacer-2) 0;
-			padding-right: var(--spacer-4);
-			text-decoration: none;
-			color: var(--color-font);
-			transition: 0.2s;
-
-			&:hover {
-				color: var(--color-secondary-dark);
-			}
-		}
-
-		&-item {
-			border-bottom: 1px solid var(--color-font);
-
-			&:first-child {
-				border-top: 1px solid var(--color-font);
-			}
-		}
-
-		&-card {
-			&-link {
-				display: flex;
-				align-items: center;
-				padding: var(--spacer-3);
-				background-color: var(--color-primary);
-				font-family: var(--font-fam-2);
-				transition: 0.2s;
-
-				// fix mouseover flicker
-				svg,
-				span {
-					pointer-events: none;
-				}
-
-				&-icon {
-					display: inline-block;
-					margin-left: auto;
-					height: 0.75rem;
-					rotate: 180deg;
-					transition: 0.3s ease-in-out;
-				}
-
-				&:hover {
-					background-color: var(--color-secondary-tint);
-
-					.tc-list-card-link-icon {
-						transform: translateX(-0.25rem);
-					}
-				}
-			}
-		}
-
-		&-cards {
-			display: grid;
-			gap: var(--spacer-1)
 		}
 	}
 }
